// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  MAHASISWA: 'MAHASISWA',
  MENTOR: 'MENTOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PackageType = {
  BIMBINGAN: 'BIMBINGAN',
  PEMBUATAN: 'PEMBUATAN'
} as const

export type PackageType = (typeof PackageType)[keyof typeof PackageType]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  EXPIRED: 'EXPIRED',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  MIDTRANS: 'MIDTRANS',
  TRANSFER_MANUAL: 'TRANSFER_MANUAL'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const ThesisStageStatus = {
  BELUM_DIMULAI: 'BELUM_DIMULAI',
  DIKERJAKAN: 'DIKERJAKAN',
  REVISI: 'REVISI',
  SELESAI: 'SELESAI'
} as const

export type ThesisStageStatus = (typeof ThesisStageStatus)[keyof typeof ThesisStageStatus]